feat: implement session board
